Seperating out oracle_specific_functions into seperate files
