+++ utils/oracle/oracle_specific_functions/enums.py
@@ -1,0 +1,11 @@
+from enum import IntEnum
+
+
+class SqlQueryValues(IntEnum):
+    S10_EVENT_STATUS = 11198
+    S19_EVENT_STATUS = 11213
+    S43_EVENT_STATUS = 11223
+    OPEN_EPISODE_STATUS_ID = 11352
+    A8_EVENT_STATUS = 11132
+    POSITIVE_APPOINTMENT_BOOKED = 11119
+    POST_INVESTIGATION_APPOINTMENT_NOT_REQUIRED = 160182
